+++ ceos/src/components/friendList/FavFriends/Favorites.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import styled, { keyframes } from "styled-components";
 import UpIc from "@assets/icons/up_ic.svg?react";
 import DowwnIc from "@assets/icons/down_ic.svg?react";
 import { SVGComponent } from "style/commonStyle";
@@ -13,36 +13,45 @@ export default function Favorites() {
   }
 
   return (
-    <Wrapper>
-      <Stroke />
+    <Wrapper isOpen={isFavListOpen}>
       <ArrangeContainer>
         <Container>
           <Text>즐겨찾기</Text>
           {isFavListOpen ? <DownIcon onClick={handleFavList} /> : <UpIcon onClick={handleFavList} />}
         </Container>
-        {isFavListOpen && <FavFriendList />}
+        <FavListContainer isOpen={isFavListOpen}>
+          {isFavListOpen && (
+            <>
+              <FavFriendList />
+            </>
+          )}
+        </FavListContainer>
       </ArrangeContainer>
-      <Stroke />
     </Wrapper>
   );
 }
 
-const Wrapper = styled.div`
+const Wrapper = styled.div<{ isOpen: boolean }>`
   display: flex;
   flex-direction: column;
-  gap: 8px;
+  align-items: center;
+  justify-content: center;
+
+  margin-top: 8px;
 `;
 
 const Container = styled.div`
   display: flex;
   justify-content: space-between;
+
   width: 343px;
+  height: 12px;
+  margin-top: 8px;
 `;
 
 const ArrangeContainer = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 8px;
 `;
 
 const FavIcon = (Component: SVGComponent) => styled(Component)`
@@ -59,8 +68,31 @@ const Text = styled.p`
   color: ${({ theme }) => theme.colors.gray500};
 `;
 
-const Stroke = styled.hr`
-  size: 0.8px;
-  width: 343px;
-  color: ${({ theme }) => theme.colors.gray200};
+const FavListContainer = styled.div<{ isOpen: boolean }>`
+  overflow: hidden;
+  min-height: ${(props) => (props.isOpen ? "68px" : 0)};
+  margin-top: 8px;
+  animation: ${(props) => (props.isOpen ? slideDown : slideUp)} 3s;
+`;
+
+const slideDown = keyframes`
+  from {
+    max-height: 0;
+    opacity: 0;
+  }
+  to {
+    max-height: 300px;
+    opacity: 1;
+  }
+`;
+
+const slideUp = keyframes`
+  from {
+    max-height: 300px;
+    opacity: 1;
+  }
+  to {
+    max-height: 0;
+    opacity: 0;
+  }
 `;
