--- ceos/src/components/message/FromBox.tsx
@@ -1,50 +1,0 @@
-import styled from "styled-components";
-import MessageTime from "./MessageTime";
-import { chatType } from "types/chatType";
-import FromMessageBox from "./FromMessageBox";
-
-interface FromBoxPropType extends chatType {
-  showProfile: boolean;
-  isSameTime: boolean;
-  lastMessage: boolean;
-  message: string;
-  time: string;
-}
-
-export default function FromBox(props: FromBoxPropType) {
-  const { user, message, time, showProfile, isSameTime, lastMessage } = props;
-
-  return (
-    <>
-      <FromWrapper>
-        {showProfile && <Name>{user}</Name>}
-        <MessageTimeWrapper $showProfile={showProfile}>
-          <FromMessageBox showProfile={showProfile} message={message} />
-          {!isSameTime || lastMessage ? <MessageTime time={time} /> : null}
-        </MessageTimeWrapper>
-      </FromWrapper>
-    </>
-  );
-}
-
-const Name = styled.p`
-  padding-left: 10px;
-  margin-top: 12px;
-  margin-bottom: 4px;
-  ${({ theme }) => theme.fonts.medium12};
-`;
-
-const FromWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const MessageTimeWrapper = styled.div<{ $showProfile: boolean }>`
-  display: flex;
-  align-items: flex-end;
-
-  margin: 0;
-  margin-bottom: ${({ $showProfile }) => ($showProfile ? "8px" : "12px")};
-  padding: 0;
-  padding-left: ${({ $showProfile }) => ($showProfile ? 0 : "50px")};
-`;
